requirements.txt
ultralytics
opencv-python
numpy
fastapi
uvicorn
python-dotenv
shapely
psycopg2-binary
